Refactoring and code improvements
